+++ functions/src/index.ts
@@ -1,9 +1,6 @@
 export {inviteMember, inviteMemberHttp} from "./inviteMember";
 export {leaveHousehold, leaveHouseholdHttp} from "./leaveHousehold";
-export {sendHouseholdInvitation, sendHouseholdInvitationHttp} from "./sendHouseholdInvitation";
 export {checkInvitation} from "./checkInvitation";
-// export {createSubscription} from "./stripe"; // Temporarily disabled due to Secret Manager requirement
 export {migrateHouseholdClaims, migrateHouseholdClaimsHttp} from "./migrateHouseholdClaims";
 export {getNutritionData} from "./nutrition";
 export {resetWeeklyUsageLimits, resetUsageLimitsNow} from "./resetUsageLimits";
-// export {createPayPalSubscription, approvePayPalSubscription} from "./paypal"; // Temporarily disabled